backend/requirements.prod.txt
fastapi>=0.135.0
uvicorn>=0.34.0
yt-dlp>=2026.1.0
python-multipart>=0.0.18
httpx>=0.28.0
langchain-openai>=1.1.0
langchain-core>=1.3.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
stripe>=8.0.0
python-dateutil>=2.8.0
pyyaml>=6.0.0
sse-starlette>=2.0.0
requests>=2.31.0
pydantic-settings>=2.0.0
openpyxl>=3.1.0
static-ffmpeg>=2.5
playwright>=1.40.0
Pillow>=10.0.0
